// src/modules/maps/maps.types.ts
export interface Area {
    id: string;
    name: string;
    country_name: string;
    country_code: string;
    administrative_division_level_1_name: string;
    administrative_division_level_1_type: string;
    administrative_division_level_2_name: string;
    administrative_division_level_2_type: string;
    administrative_division_level_3_name: string;
    administrative_division_level_3_type: string;
    postal_code: number;
}

export interface SearchAreaResponse {
    success: boolean;
    areas: Area[];
}